forge/policy-editor/python/pyproject.toml
[project]
name = "rfnry-example-policy-editor"
version = "0.0.1"
description = "Policy editor — verified document editing for HR/compliance teams (rfnry + forge-scribe)"
requires-python = ">=3.12"
dependencies = [
    "rfnry",
    "rfnry-forge-scribe[rfnry]",
    "anthropic>=0.40",
    "fastapi>=0.136.0",
    "uvicorn[standard]>=0.44.0",
    "pydantic>=2.13.2",
]

[tool.uv.sources]
rfnry = { path = "../../../../../rfnry/packages/python", editable = true }
rfnry-forge-scribe = { path = "../../../../../forge/packages/scribe", editable = true }
rfnry-forge-parser = { path = "../../../../../forge/packages/parser", editable = true }

[project.optional-dependencies]
dev = [
    "mypy>=1.20.1",
    "poethepoet>=0.44.0",
    "ruff>=0.15.11",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.mypy]
python_version = "3.12"

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.ruff.format]
quote-style = "double"

[tool.poe.tasks]
dev = "python -m src.main"
start = "python -m src.main"
format = "ruff format ."
"format:imports" = "ruff check --select I --fix ."
check = "ruff check ."
"check:fix" = "ruff check --fix ."
typecheck = "mypy src"
